+++ src/config/imap.php
@@ -37,6 +37,32 @@ return [
 
     /*
     |--------------------------------------------------------------------------
+    | Security options
+    |--------------------------------------------------------------------------
+    |
+    | You can enable or disable certain security features here by setting them to true or false to enable or disable
+    | them.
+    | -detect_spoofing:
+    |       Detect spoofing attempts by checking the message sender against the message headers.
+    |       Default TRUE
+    | -detect_spoofing_exception:
+    |       Throw an exception if a spoofing attempt is detected.
+    |       Default FALSE
+    | -sanitize_filenames:
+    |       Sanitize attachment filenames by removing any unwanted and potentially dangerous characters. This is not a
+    |       100% secure solution, but it should help to prevent some common attacks. Please sanitize the filenames
+    |       again if you need a more secure solution.
+    |       Default TRUE
+    |
+    */
+    'security' => [
+        "detect_spoofing" => true,
+        "detect_spoofing_exception" => false,
+        "sanitize_filenames" => true,
+    ],
+
+    /*
+    |--------------------------------------------------------------------------
     | Available accounts
     |--------------------------------------------------------------------------
     |