+++ src/Query/Query.php
@@ -164,7 +164,10 @@ class Query {
                 if ($statement[1] === null) {
                     $query .= $statement[0];
                 } else {
-                    if (is_numeric($statement[1])) {
+                    if (is_numeric($statement[1]) || (
+                            ($statement[0] === 'SINCE' || $statement[0] === 'BEFORE') &&
+                            $this->client->getConfig()->get('options.unescaped_search_dates', false)
+                    )) {
                         $query .= $statement[0] . ' ' . $statement[1];
                     } else {
                         $query .= $statement[0] . ' "' . $statement[1] . '"';