+++ CHANGELOG.md
@@ -14,6 +14,7 @@ Updates should follow the [Keep a CHANGELOG](http://keepachangelog.com/) princip
 - Security configuration options added
 - Spoofing detection added #40
 - RFC4315 MOVE fallback added #123 (thanks @freescout-help-desk)
+- Content fetching RFC standard support added #510 (thanks @ybizeul)
 
 ### Breaking changes
 - NaN